Enhancement/70 contact page layout (#87)
* Add template API endpoint and API parsing tool

A new api endpoint was added to account for new 'template' models
in Contentful. This allows the use of predefined page templates
which can be populated with the desired data.

Additionally, a tool to parse Contentful API responses was added
to reduce (and possibly eliminate) the previous dependence on special
'API/' components whose sole purpose was to parse API response
data into a format useable by the UI components.

The new parsing tool takes a typical Contentful API response as input
and outputs a data object with a 'type' key, representing the content
model name, and a 'fields' key, which contains fully parsed and
resolved Asset and Entry link fields (with the notable caveat being
circular links are only resolved to one level deep to avoid infinite
recursion). This tool also supports pagination when a 'limit' parameter
is specified.

* Add contact template and components, update AWS config

* Update brand page to use api parsing tool

* Update navigation to use api parsing tool

* Convert px values to rem